tests/wpt/tests/html/semantics/embedded-content/media-elements/track/track-element/track-webvtt-cue-size-align.html

<!DOCTYPE html>
<title>Cue size and alignment from settings</title>
<script src="track-helpers.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
check_cues_from_track("resources/cue-size-align.vtt", function(track) {
    var expected = [
        { size: 100, align: "start"  },
        { size: 10,  align: "end"    },
        { size: 0,   align: "center" }
    ];

    assert_cues_match(track.cues, expected);
});

check_cues_from_track("resources/cue-size-align-bad.vtt", function(track) {
    var expected = [
        { size: 100, align: "center" },
        { size: 100, align: "end"    },
        { size: 100, align: "center" }
    ];

    assert_cues_match(track.cues, expected);
});
</script>
